mysite/requirements.txt
attrs==19.1.0
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
clarifai==2.6.2
configparser==3.8.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-ckeditor==5.8.0
django-js-asset==1.2.2
docutils==0.15.2
EasyProcess==0.2.7
emoji==0.5.3
fake-useragent==0.1.11
future==0.17.1
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
jsmin==2.2.2
jsonschema==3.0.2
lxml==4.4.1
Markdown==2.6.11
MeaningCloud-python==1.1.1
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
Pillow==6.2.1
pipenv==2018.11.26
plyer==1.4.0
protobuf==3.9.1
psycopg2==2.8.3
psycopg2-binary==2.8.4
PyJWT==1.7.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
slugify==0.0.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
soupsieve==1.9.3
sqlparse==0.3.0
tqdm==4.35.0
urllib3==1.25.3
virtualenv==16.7.5
virtualenv-clone==0.5.3
whitenoise==5.0.1